move key connector components to KM team ownership
