+++ apps/desktop/src/key-management/key-connector/remove-password.component.html
