+++ libs/key-management-ui/src/index.ts
@@ -4,3 +4,4 @@
 
 export { LockComponent } from "./lock/components/lock.component";
 export { LockComponentService, UnlockOptions } from "./lock/services/lock-component.service";
+export { RemovePasswordComponent } from "./key-connector/remove-password.component";
